rename RareMatrix to SparseMatrix
