--- sem_2/RareMatrix/readme.txt
@@ -1,2 +1,0 @@
-Program that works with rare matrix - a matrix with great number of zeros - O(n^2).
-(c) rdmr, 2005
