+++ sem_2/SparseMatrix/readme.txt
@@ -1,0 +1,2 @@
+Program that works with sparse matrix - a matrix with great number of zeros - O(n^2).
+(c) rdmr, 2005
